pyproject.toml
[project]
name = "lang-chain-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-anthropic>=0.3.13",
    "python-dotenv>=1.1.0",
]
